refactor(go/plugins/a2ui): move the plugin to a2ui/exp while in preview
A2UI support is still in preview, so it now lives at
go/plugins/a2ui/exp as package exp, matching plugins/middleware/exp
and plugins/firebase/exp. The plugin name stays "a2ui": there is no
stable a2ui plugin to collide with, so the Dev UI name, the
a2ui-catalog value type, and parity with the JS plugin are unchanged.

Callers import it as a2uix, following the aix/genkitx/middlewarex
convention.
