+++ go/plugins/a2ui/exp/part.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package a2ui
+package exp
 
 import "github.com/firebase/genkit/go/ai"
